BACKEND/requirement.txt
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
cached-property==1.5.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
comm==0.2.1

cryptography==3.4.8

curlify==2.2.1
DateTime==5.4
dbus-python==1.2.18
debugpy==1.8.0
decorator==4.4.2

defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0

Django==4.1.4
django-cors-headers==3.13.0
django-postgresql==0.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
docutils==0.17.1

exceptiongroup==1.2.0
executing==2.0.1
facebook-business==18.0.5
fasteners==0.14.1
filelock==3.6.0
freezegun==1.0.0
frozenlist==1.4.1
future==0.18.2
geoip2==2.9.0
gevent==21.8.0
greenlet==1.1.2
gunicorn==21.2.0

html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.0
ipython==8.20.0
isodate==0.6.1
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==23.5.0

launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libsass==0.20.1
lockfile==0.12.2


mysqlclient==2.2.0


numpy==1.26.3



Pillow==9.0.1

psycopg2==2.9.2

PyJWT==2.8.0

pyparsing==2.4.7



pytz==2022.7



requests==2.25.1

six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3

sqlparse==0.4.3














